backend/requirements.txt
# Core dependencies
fastapi>=0.104.0
uvicorn>=0.23.2
python-dotenv>=1.0.0
pydantic>=2.4.2
pydantic-settings>=2.0.3
fastapi[standard]

# Database
sqlalchemy>=2.0.21
alembic>=1.12.0
psycopg2-binary>=2.9.9
asyncpg>=0.28.0

# Authentication & Security
python-jose>=3.3.0
passlib>=1.7.4
python-multipart>=0.0.6
bcrypt>=4.0.1

# Bright Data integration
requests>=2.31.0
aiohttp>=3.8.5

# AI/ML
litellm>=0.6.0
openai>=1.2.0
google-adk

# Utilities
python-dateutil>=2.8.2
tenacity>=8.2.3
loguru>=0.7.2
